Add 'completions' command
We establish the readline-equivalent $input word for consistency, so
that the $input value that we pass to processCallback() will be
equivalent to the value passed by GNU Readline in the same situation.

This is not strictly necessary for the present code, but may prove to
be useful for subsequent enhancements.